+++ src/Client/ClientSetup.h
@@ -178,6 +178,7 @@ public:
         bypass->modules.push_back(new Module("Transition Customiser", "custom-trans", "Adjust your scene transition"));
 
         bypass->modules.push_back(new Module("Pitch Shifter", "pitch-shifter", "Shift the pitch of the game audio"));
+        bypass->modules.push_back(new Module("No Short Numbers", "no-short-nums", "Show the full value instead of the short value in level cells"));
         
         Client::instance->windows.push_back(bypass);